# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MECHA API",
        "description": "API untuk Sistem Manajemen Bengkel Mobil",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/bengkels": {
            "get": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Get semua bengkel",
                "operationId": "798227e7b9a4af2a632de95f2cd9067a",
                "responses": {
                    "200": {
                        "description": "Data bengkels retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Tambah bengkel baru",
                "operationId": "7f03be5f366260227f600f9f639ef025",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone",
                                    "email",
                                    "city",
                                    "province",
                                    "status",
                                    "owner_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    },
                                    "owner_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bengkel created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/bengkels/{id}": {
            "get": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Get bengkel by ID",
                "operationId": "d0e9e0d852ff4605acab68a600340e5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bengkel retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Update bengkel",
                "operationId": "24cc27ecab821c2a5e064206c907120c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bengkel updated successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Hapus bengkel",
                "operationId": "cd6472412d29e9a4ea579c0d3eded0d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bengkel deleted successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/diagnostics": {
            "get": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Get semua diagnostics",
                "operationId": "6e7b767bc65c694579ba7584217e0c48",
                "responses": {
                    "200": {
                        "description": "Data diagnostics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Tambah diagnostic baru",
                "operationId": "01615aeab16e9eadc3989c019c43e758",
                "responses": {
                    "201": {
                        "description": "Diagnostic created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/diagnostics/{id}": {
            "get": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Get diagnostic by ID",
                "operationId": "a73d84a382b0fe5903d8b68dea38c85c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diagnostic retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Update diagnostic",
                "operationId": "2ed4897c68c00406df6d356abc29736d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diagnostic updated successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Hapus diagnostic",
                "operationId": "365b6f7df30ea6540bf18cc79781a9ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Diagnostic deleted successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}